Move common test code
